🐕 Husky and lint-staged
Husky is a tool that allows you to run scripts when you commit or push your code. It is a great way to run tests and linting before you push your code to the repository.

lint-staged is a tool that allows you to run scripts on only the files that have been staged for commit. This is a great way to run tests and linting on only the files that have changed.

📦 Installation
```bash
npx mrm@2 lint-staged
```